# Optimize MiniCPMO mask creation with vectorized implementation