@tc-mb 2025-08-07
> cc @tc-mb can you verify this?

Of course, I will look into verifying the optimization and helping with this PR.