@tc-mb 2025-08-08
@DarkLight1337  I ran a side-by-side test, and the results are correct. This is a good optimization.

@skyloevil  Could I incorporate this optimization into other code implementations? I'll include a link to this pull request and your account name as a reference.